Initial commit of Life Church backup scripts
These backup scripts will be used to automatically maintain backups of
the data on LifeNAS, our primary file server. The data is backed up to
another NAS on the same network using rsync and loop devices across a
CIFS share.

This script will be used to mount the various volumes needed by the main
backup script if they're not already mounted.